fix logic for intializing openai clients
